Renamed Register to RegisterMethods
